移除不必要的引用

+++ src/main/java/com/upec/factoryscheduling/aps/service/SchedulingService.java
@@ -129,14 +129,12 @@ public class SchedulingService {
         SolverJob<FactorySchedulingSolution, Long> solverJob = solverManager.solveAndListen(
                 problemId,  // 问题标识
                 id -> problem,  // 提供问题数据的函数
-
                 // 每次找到更好的解决方案时的回调函数
                 solution -> {
                     // 记录新的最佳解决方案分数
                     log.info("New best solution found: {}", solution.getScore());
                     // 此处可以扩展，例如更新UI或临时保存中间结果
                 },
-
                 // 求解完成时的回调函数
                 finalBestSolution -> {
                     // 记录最终最佳解决方案分数
@@ -144,7 +142,6 @@ public class SchedulingService {
                     // 保存最终调度结果到数据库
                     saveSolution(finalBestSolution);
                 },
-
                 // 求解出错时的回调函数
                 (id, throwable) -> {
                     log.error("Scheduling error: {}", throwable.getMessage());
@@ -589,7 +586,7 @@ public class SchedulingService {
 
             // 计算当日累计使用工时
             double countDailyHours = timeslotList.stream()
-                    .mapToDouble(t -> t.getDuration().doubleValue())
+                    .mapToDouble(t -> t.getDuration())
                     .sum();
 
             // 获取时间和设备信息
@@ -615,7 +612,7 @@ public class SchedulingService {
             long overlapTime = timeslotList.stream()
                     .mapToLong(t -> DateTimeCalculatorUtil.overlapTime(
                             timeslot.getStartTime().getMinute(),
-                            timeslot.getStartTime().plusMinutes((int) (timeslot.getDuration().doubleValue() * 60)).getMinute(),
+                            timeslot.getStartTime().plusMinutes((int) (timeslot.getDuration() * 60)).getMinute(),
                             t.getStartTime().getMinute(),
                             t.getStartTime().plusMinutes(t.getStartTime().getMinute()).getMinute()))
                     .sum();
